+++ samples/kotlin/src/main/AndroidManifest.xml
@@ -47,7 +47,7 @@
         </receiver>
 
         <activity
-            android:name="com.bingo.foo.MainActivity"
+            android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
