+++ samples/kotlin/src/main/java/com/forgerock/kotlinapp/UserInfoFragment.kt
@@ -9,7 +9,6 @@ import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
 import androidx.fragment.app.Fragment
-import com.bingo.foo.ActivityListener
 
 
 // TODO: Rename parameter arguments, choose names that match
