+++ samples/kotlin/src/main/res/layout/fragment_user_info.xml
@@ -67,7 +67,7 @@
         android:layout_marginEnd="8dp"
         android:height="100dp"
         android:scrollbars = "vertical"
-        android:textColor="@color/purple_500"
+        android:textColor="@color/black"
         android:gravity="center"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView3"
@@ -95,7 +95,7 @@
         android:layout_marginEnd="8dp"
         android:height="100dp"
         android:scrollbars = "vertical"
-        android:textColor="@color/material_dynamic_primary70"
+        android:textColor="@color/black"
         android:gravity="center"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView4"
